update the documentation.

+++ src/pythagorean.c
@@ -2,7 +2,7 @@
 #include "square_root.h"
 
 
-float Pythagorean_hypotenuse(int x, int y)
+float Pythagorean_hypotenuse(float x, float y)
 {
     float hypotenuse = SquareRoot_sqrt(x*x + y*y);
     return hypotenuse;
