+++ src/pythagorean.h
@@ -2,7 +2,7 @@
 #define PYTHAGOREAN_H
 
 
-float Pythagorean_hypotenuse(int x, int y);
+float Pythagorean_hypotenuse(float x, float y);
 
 
 #endif //PYTHAGOREAN_H
